+++ NEWS.md
@@ -5,8 +5,10 @@
 ### New APIs versions
 * Provides `API_NAME vX.Y`
 * Requires `API_NAME vX.Y`
+* Provides `marc-specifications` `v1.1`
 
 ### Features
+* Update marc specification to support allowed values for 008 field positions ([MODQM-405](https://github.com/folio-org/mod-quick-marc/pull/244))
 * support new Authority record creation ([MODQM-322](https://issues.folio.org/browse/MODQM-322)
 
 ### Bug fixes